Add story tracking microservice: implement article retrieval and matching modules
